server/requirements.txt
ffs
git+git://github.com/wpjunior/django-mongotools.git
django-templated-email
-e git://github.com/pinax/django-mailer.git#egg=django-mailer
django-celery-email
django-celery
celery-with-redis
django-tastypie
git+git://github.com/jacobian/django-pjax.git
django-tastypie-mongoengine
django-crispy-forms

